requirements.txt
djangorestframework 
django-cors-headers
django_celery_beat
django_celery_results
python-dotenv
django-ratelimit
sentry-sdk
openai
djangorestframework-simplejwt
requests